# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "欺骗防御接口文档",
        "contact": {
            "name": "akita.tian",
            "email": "chronoyq@163.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "2.0"
    },
    "host": "127.0.0.1:8082",
    "paths": {
...
        "/api/v1/honeypot": {
            "post": {
                "description": "创建蜜罐",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "蜜罐管理"
                ],
                "summary": "创建蜜罐",
                "parameters": [
                    {
                        "description": "HoneypotName",
                        "name": "HoneypotName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Honeypot"
                        }
                    },
                    {
                        "description": "ImageAddress",
                        "name": "ImageAddress",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Honeypot"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"内部异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "5001": {
                        "description": "{\"code\":5001,\"msg\":\"蜜罐服务器不存在、请检测蜜罐服务状态\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "5002": {
                        "description": "{\"code\":5002,\"msg\":\"K8S创建镜像失败\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "5006": {
                        "description": "{\"code\":5006,\"msg\":\"蜜罐名称存在\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "7001": {
                        "description": "{\"code\":7001,\"msg\":\"镜像端口或是服务未配置、请先进行配置\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/honeypot/:id": {
            "get": {
                "description": "获得蜜罐详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "蜜罐管理"
                ],
                "summary": "获得蜜罐详细信息",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "int valid",
                        "name": "int",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"内部异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "5005": {
                        "description": "{\"code\":5005,\"msg\":\"K8S删除蜜罐错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除蜜罐接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "蜜罐管理"
                ],
                "summary": "删除蜜罐接口",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "int valid",
                        "name": "int",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"内部异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "5005": {
                        "description": "{\"code\":5005,\"msg\":\"K8S删除蜜罐错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/honeypot/protocol": {
            "get": {
                "description": "查看已经进行协议转发的蜜罐",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "蜜罐管理"
                ],
                "summary": "查看已经进行协议转发的蜜罐",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"请求参数错误\",\"data\":[\"honeypot-10.42.0.41\"]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/honeypot/set": {
            "post": {
                "description": "查找蜜罐接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "蜜罐管理"
                ],
                "summary": "查找蜜罐接口",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "Payload",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/comm.HoneypotSelectPayload"
                        }
                    },
                    {
                        "description": "ProtocolType",
                        "name": "ProtocolType",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/comm.HoneypotSelectPayload"
                        }
                    },
                    {
                        "description": "PageNumber",
                        "name": "PageNumber",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comm.HoneypotSelectPayload"
                        }
                    },
                    {
                        "description": "PageSize",
                        "name": "PageSize",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comm.HoneypotSelectPayload"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comm.HoneypotSelectResultPayload"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"内部异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "comm.HoneypotSelectPayload": {
            "type": "object",
            "required": [
                "PageNumber",
                "PageSize"
            ],
            "properties": {
                "PageNumber": {
                    "description": "页number",
                    "type": "integer"
                },
                "PageSize": {
                    "description": "页size",
                    "type": "integer"
                },
                "Payload": {
                    "description": "查找payload",
                    "type": "string"
                },
                "ProtocolType": {
                    "description": "协议类型",
                    "type": "string"
                }
            }
        },
        "comm.HoneypotSelectResultPayload": {
            "type": "object",
            "properties": {
                "CreateTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "Creator": {
                    "description": "创建用户",
                    "type": "string"
                },
                "HoneypotIP": {
                    "description": "蜜罐IP",
                    "type": "string"
                },
                "HoneypotName": {
                    "description": "蜜罐名称",
                    "type": "string"
                },
                "ID": {
                    "description": "蜜罐ID",
                    "type": "integer"
                },
                "ServerIP": {
                    "description": "蜜网IP",
                    "type": "string"
                },
                "ServerType": {
                    "description": "蜜罐类型",
                    "type": "string"
                },
                "Status": {
                    "description": "蜜罐状态",
                    "type": "integer"
                }
            }
        },
...
        "models.Honeypot": {
            "type": "object",
            "required": [
                "HoneypotName",
                "ImageAddress"
            ],
            "properties": {
                "CreateTime": {
                    "type": "string"
                },
                "Creator": {
                    "type": "string"
                },
                "HoneypotIP": {
                    "type": "string"
                },
                "HoneypotName": {
                    "description": "蜜罐名称",
                    "type": "string"
                },
                "ImageAddress": {
                    "description": "镜像地址",
                    "type": "string"
                },
                "PodName": {
                    "type": "string"
                },
                "ServerPort": {
                    "type": "integer"
                },
                "ServerType": {
                    "type": "string"
                },
                "ServersID": {
                    "type": "integer"
                },
                "Status": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "servers": {
                    "$ref": "#/definitions/models.HoneypotServers"
                }
            }
        },
        "models.HoneypotServers": {
            "type": "object",
            "properties": {
                "AgentID": {
                    "description": "代理ID",
                    "type": "string"
                },
                "CreateTime": {
                    "description": "注册时间",
                    "type": "string"
                },
                "HeartbeatTime": {
                    "description": "心跳时间",
                    "type": "string"
                },
                "HostName": {
                    "description": "主机名称",
                    "type": "string"
                },
                "ServerIP": {
                    "description": "蜜网IP",
                    "type": "string"
                },
                "Status": {
                    "description": "状态",
                    "type": "integer"
                },
                "id": {
                    "description": "蜜网ID",
                    "type": "integer"
                }
            }
        },
...
    }
}